# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Steam Clone API
  version: 0.1.0
  description: Hexagonal backend for Steam-like platform
servers:
  - url: http://localhost:8080
paths:
  /auth/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: User registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /auth/login:
    post:
      summary: Login with credentials
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /auth/me:
    get:
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
  /auth/logout:
    post:
      responses:
        '204':
          description: Logged out
  /users/{id}:
    get:
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: User detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
    put:
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
  /users/{id}/library:
    get:
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Library entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LibraryEntryResponse'
  /games:
    get:
      responses:
        '200':
          description: Game list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GameResponse'
  /games/{id}:
    get:
      parameters:
        - $ref: '#/components/parameters/GameId'
      responses:
        '200':
          description: Game detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameResponse'
  /games/{id}/reviews:
    get:
      parameters:
        - $ref: '#/components/parameters/GameId'
      responses:
        '200':
          description: Reviews for game
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReviewResponse'
  /reviews:
    post:
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReviewRequest'
      responses:
        '200':
          description: Created review
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewResponse'
  /reviews/{id}:
    put:
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ReviewId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReviewRequest'
      responses:
        '200':
          description: Updated review
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewResponse'
    delete:
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ReviewId'
      responses:
        '204':
          description: Review removed
  /recommendations/{userId}:
    get:
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Recommended games
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecommendationResponse'
  /genres:
    get:
      responses:
        '200':
          description: Genre catalog
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GenreResponse'
  /platforms:
    get:
      responses:
        '200':
          description: Platform catalog
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PlatformResponse'
  /library/add:
    post:
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LibraryEntryRequest'
      responses:
        '200':
          description: Added entry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LibraryEntryResponse'
  /library/remove/{gameId}:
    delete:
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GameId'
      responses:
        '204':
          description: Removed entry
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UserId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    GameId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ReviewId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
    RegisterRequest:
      type: object
      required: [email, username, password]
      properties:
        email:
          type: string
          format: email
        username:
          type: string
        password:
          type: string
          format: password
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    AuthResponse:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
        roles:
          type: array
          items:
            type: string
        accessToken:
          type: string
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        username:
          type: string
        roles:
          type: array
          items:
            type: string
        favoriteGenres:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
    UpdateUserRequest:
      type: object
      properties:
        username:
          type: string
        favoriteGenres:
          type: array
          items:
            type: string
    GameResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        basePrice:
          $ref: '#/components/schemas/Money'
        currentPrice:
          $ref: '#/components/schemas/Money'
        releaseDate:
          type: string
          format: date-time
        genres:
          type: array
          items:
            type: string
        platforms:
          type: array
          items:
            type: string
    Money:
      type: object
      properties:
        amount:
          type: number
          format: double
        currency:
          type: string
    ReviewResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        gameId:
          type: string
          format: uuid
        rating:
          type: integer
        comment:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateReviewRequest:
      type: object
      required: [gameId, rating, comment]
      properties:
        gameId:
          type: string
          format: uuid
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
    UpdateReviewRequest:
      type: object
      properties:
        rating:
          type: integer
        comment:
          type: string
    RecommendationResponse:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        games:
          type: array
          items:
            $ref: '#/components/schemas/GameResponse'
    GenreResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
    PlatformResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
    LibraryEntryRequest:
      type: object
      properties:
        gameId:
          type: string
          format: uuid
        statuses:
          type: array
          items:
            type: string
        progressMinutes:
          type: integer
    LibraryEntryResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        gameId:
          type: string
          format: uuid
        statuses:
          type: array
          items:
            type: string
        progressMinutes:
          type: integer
        addedAt:
          type: string
          format: date-time