# backend/src/main/java/com/alidev/steamclone/infrastructure/rest/AuthController.java
package com.alidev.steamclone.infrastructure.rest;

import com.alidev.steamclone.application.dto.auth.AuthResponse;
import com.alidev.steamclone.application.dto.auth.LoginRequest;
import com.alidev.steamclone.application.dto.auth.RegisterRequest;
import com.alidev.steamclone.application.dto.user.UserResponse;
import com.alidev.steamclone.application.services.AuthService;
import com.alidev.steamclone.application.services.UserService;
import com.alidev.steamclone.infrastructure.security.SecurityUtils;
import jakarta.validation.Valid;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/auth")
public class AuthController {

    private final AuthService authService;
    private final UserService userService;

    public AuthController(AuthService authService, UserService userService) {
        this.authService = authService;
        this.userService = userService;
    }

    @PostMapping("/register")
    public ResponseEntity<AuthResponse> register(@Valid @RequestBody RegisterRequest request) {
        return ResponseEntity.ok(authService.register(request));
    }

    @PostMapping("/login")
    public ResponseEntity<AuthResponse> login(@Valid @RequestBody LoginRequest request) {
        return ResponseEntity.ok(authService.login(request));
    }

    @GetMapping("/me")
    public ResponseEntity<UserResponse> me() {
        return ResponseEntity.ok(userService.getById(SecurityUtils.currentUserId()));
    }

    @PostMapping("/logout")
    public ResponseEntity<Void> logout() {
        return ResponseEntity.noContent().build();
    }
}